+++ loki/loki.d/local/share/cook/bin/configure-syslog-ng.sh
@@ -24,9 +24,6 @@ config_version=$(/usr/local/sbin/syslog-ng --version | \
   sed "s${sep}%%myip%%${sep}$IP${sep}g" \
   > /usr/local/etc/syslog-ng.conf
 
-# create remote log dir
-mkdir -p /mnt/log/remote/metrics
-
 # stop and disable syslogd
 service syslogd onestop || true
 service syslogd disable