+++ loki/loki.sh
@@ -124,32 +124,11 @@ pkg install -y vault
 step "Install package syslog-ng"
 pkg install -y syslog-ng
 
+step "Install package grafana-loki"
+pkg install -y grafana-loki
+
 step "Clean package installation"
 pkg clean -y
-
-# last updated 2022-07-20
-step "Download loki release from github"
-fetch -qo - https://github.com/grafana/loki/releases/download/\
-v2.6.1/loki-freebsd-amd64.zip | unzip -p - loki-freebsd-amd64 \
-  >/usr/local/bin/loki
-chmod 755 /usr/local/bin/loki
-
-if [ "$(sha256 -q /usr/local/bin/loki)" != \
-  "ece8b6d91fcf9a3b0cae6f828988c23cb077e17662fff4e5e0066570d29aa3d5" ]; then
-  exit_error "/usr/local/bin/loki checksum mismatch!"
-fi
-
-# last updated 2022-07-20
-step "Download promtail release from github"
-fetch -qo - https://github.com/grafana/loki/releases/download/\
-v2.6.1/promtail-freebsd-amd64.zip | unzip -p - promtail-freebsd-amd64 \
-  >/usr/local/bin/promtail
-chmod 755 /usr/local/bin/promtail
-
-if [ "$(sha256 -q /usr/local/bin/promtail)" != \
-  "7aa244fb3d0f8dba8b9edff7536fa55be2300bee7c63b0b003ac1a8f7d99ab9d" ]; then
-  exit_error "/usr/local/bin/promtail checksum mismatch!"
-fi
 
 # -------------- END PACKAGE SETUP -------------